add data validation

+++ backend/Controllers/CommentController.cs
@@ -18,13 +18,15 @@ namespace backend.Controllers
 
         [HttpGet]
         public async Task<IActionResult> GetAll() {
+            if(!ModelState.IsValid) return BadRequest(ModelState);
             var comments = await _commentRepo.GetAllAsync();
             var commentDto = comments.Select(s => s.ToCommentDto());
             return Ok(commentDto);
         }
 
-        [HttpGet("{id}")]
+        [HttpGet("{id:int}")]
         public async Task<IActionResult> GetById([FromRoute] int id) {
+            if(!ModelState.IsValid) return BadRequest(ModelState);
             var comment = await _commentRepo.GetByIdAsync(id);
             if(comment == null) {
                 return NotFound();
@@ -32,19 +34,33 @@ namespace backend.Controllers
             return Ok(comment.ToCommentDto());
         }
 
-        [HttpPost("{stockId}")]
+        [HttpPost("{stockId:int}")]
         public async Task<IActionResult> Create([FromRoute] int stockId, CreateCommentDto createCommentDto) {
+            if(!ModelState.IsValid) return BadRequest(ModelState);
             if(!await _stockRepo.StockExists(stockId)) {
                 return BadRequest("Stock does not exist");
             }
             var commentModel = createCommentDto.ToCommentFromCreate(stockId);
             await _commentRepo.CreateAsync(commentModel);
-            return CreatedAtAction(nameof(GetById), new { id = commentModel}, commentModel.ToCommentDto());
+            return CreatedAtAction(nameof(GetById), new { id = commentModel.Id}, commentModel.ToCommentDto());
         }
 
+        [HttpPut]
+        [Route("{id:int}")]
+        public async Task<IActionResult> Update([FromRoute] int id, [FromBody] UpdateCommentRequestDto updateDto) {
+            if(!ModelState.IsValid) return BadRequest(ModelState);
+        var commentModel = await _commentRepo.UpdateAsync(id, updateDto.ToCommentFromUpdate(id));
+
+        if(commentModel == null) {
+            return NotFound("Comment not found");
+        }
+        return Ok(commentModel.ToCommentDto());
+       }
+
         [HttpDelete]
-        [Route("{id}")]
+        [Route("{id:int}")]
         public async Task<IActionResult> Delete([FromRoute] int id) {
+            if(!ModelState.IsValid) return BadRequest(ModelState);
             var commentModel = await _commentRepo.DeleteAsync(id);
             if(commentModel == null) 
             {
